kernel: move PanicWriter to platform/chip
