+++ kernel/src/platform/chip.rs
@@ -6,6 +6,7 @@
 
 use crate::platform::mpu;
 use crate::syscall;
+use crate::utilities::io_write::IoWrite;
 use core::fmt::Write;
 
 /// Interface for individual MCUs.
@@ -176,3 +177,23 @@ impl ClockInterface for NoClockControl {
 /// Instance of NoClockControl for things that need references to
 /// `ClockInterface` objects.
 pub const NO_CLOCK_CONTROL: NoClockControl = NoClockControl {};
+
+/// Interface for chips to create a synchronous writer for panics.
+///
+/// Any mechanism that can output a panic message during a panic must implement
+/// [`PanicWriter`] to enable the `panic()` functions to write the output. This
+/// requires the mechanism to provide a new constructor for the writer that
+/// creates a synchronous writer that implements [`IoWrite`].
+pub trait PanicWriter {
+    /// The configuration data the mechanism needs to configure the writer for
+    /// panic output.
+    type Config;
+
+    /// Create a new synchronous writer capable of sending panic messages.
+    ///
+    /// The constructed writer must be created on the stack. Because panic
+    /// will never return this is effectively a static allocation.
+    ///
+    /// The writer must implement [`IoWrite`].
+    unsafe fn create_panic_writer(config: Self::Config) -> impl IoWrite;
+}
